chore: refactored document creation backend code; fixed integration test
