--- src/main/java/net/atos/zac/app/informatieobjecten/model/RestDocumentCreationData.kt
@@ -1,15 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 Atos
- * SPDX-License-Identifier: EUPL-1.2+
- */
-
-package net.atos.zac.app.informatieobjecten.model;
-
-import java.util.UUID;
-
-public class RestDocumentCreationData {
-
-    public UUID zaakUUID;
-
-    public String taskId;
-}
