+++ src/main/kotlin/net/atos/zac/app/documentcreation/DocumentCreationRestService.kt
@@ -7,6 +7,7 @@ package net.atos.zac.app.documentcreation
 
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
+import jakarta.validation.Valid
 import jakarta.ws.rs.Consumes
 import jakarta.ws.rs.POST
 import jakarta.ws.rs.Path
@@ -14,8 +15,8 @@ import jakarta.ws.rs.Produces
 import jakarta.ws.rs.core.MediaType
 import net.atos.client.zgw.zrc.ZrcClientService
 import net.atos.client.zgw.ztc.ZtcClientService
-import net.atos.zac.app.informatieobjecten.model.RestDocumentCreationData
-import net.atos.zac.app.informatieobjecten.model.RestDocumentCreationResponse
+import net.atos.zac.app.documentcreation.model.RestDocumentCreationData
+import net.atos.zac.app.documentcreation.model.RestDocumentCreationResponse
 import net.atos.zac.app.util.exception.InputValidationFailedException
 import net.atos.zac.configuratie.ConfiguratieService
 import net.atos.zac.documentcreation.SmartDocumentsService
@@ -40,7 +41,9 @@ class DocumentCreationRestService @Inject constructor(
 
     @POST
     @Path("/createdocumentattended")
-    fun createDocumentAttended(restDocumentCreationData: RestDocumentCreationData): RestDocumentCreationResponse {
+    fun createDocumentAttended(
+        @Valid restDocumentCreationData: RestDocumentCreationData
+    ): RestDocumentCreationResponse {
         val zaak = zrcClientService.readZaak(restDocumentCreationData.zaakUUID)
         assertPolicy(policyService.readZaakRechten(zaak).creeerenDocument)
