+++ src/main/kotlin/net/atos/zac/app/documentcreation/model/RestDocumentCreationData.kt
@@ -1,0 +1,17 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Lifely
+ * SPDX-License-Identifier: EUPL-1.2+
+ */
+package net.atos.zac.app.documentcreation.model
+
+import jakarta.validation.constraints.NotNull
+import nl.lifely.zac.util.NoArgConstructor
+import java.util.UUID
+
+@NoArgConstructor
+class RestDocumentCreationData(
+    @field:NotNull
+    var zaakUUID: UUID,
+
+    var taskId: String? = null
+)
